some linux

+++ android/install_ndk.py
@@ -1,25 +1,25 @@
 
 import os
 import sys
+import glob
 import shutil
 import platform
 import urllib.request
 from pathlib import Path
 from enum import Enum
 
-if Path("ndk").exists:
-    print("NDK exists")
 
 is_windows = platform.system() == "Windows"
 is_mac     = platform.system() == "Darwin"
 is_linux   = platform.system() == "Linux"
+
+is_unix = is_mac or is_linux
 
 
 def run(command: str):
     print(command)
     if os.system(command):
         raise Exception("Shell script has failed")
-
 
 
 class Arch(Enum):
@@ -38,16 +38,27 @@ class Arch(Enum):
 
 
 def setup():
-    shutil.rmtree("ndk")
+    if Path("ndk").exists():
+        print("NDK exists")
+        shutil.rmtree("ndk")
+
     if is_mac:
         run("brew install p7zip")
 
-
+def python_exe() -> str:
+    if is_windows:
+        return "py"
+    else:
+        return "python3"
+        
 def download_link():
     link = "https://dl.google.com/android/repository/android-ndk-r25c-darwin.dmg"
 
     if is_windows:
         link = "https://dl.google.com/android/repository/android-ndk-r25c-windows.zip"
+
+    if is_linux:
+        link = "https://dl.google.com/android/repository/android-ndk-r25c-linux.zip"
 
     print("Download link: " + link)
     return link
@@ -66,27 +77,31 @@ def download() -> str:
 
 def make_toolchain(make_tool: str, arch: Arch, api_level: str = "21"):
     root = f"ndk/{arch}"
-    run(f"py {make_tool} --api {api_level} --arch {arch} --install-dir {root}")
+    run(f"{python_exe()} {make_tool} --api {api_level} --arch {arch} --install-dir {root}")
     ndk_bin = os.path.abspath(root) + "/bin"
     os.environ["PATH"] += os.pathsep + ndk_bin
+
+    if not is_windows:
+        for file in glob.glob(ndk_bin + "/*"):
+            run("chmod +x " + file)
 
     clang = f"{ndk_bin}/{arch.clang()}"
 
     print(f"Clang path:")
     print(clang)
 
-    os.remove(clang)
+    if is_windows:
+        os.remove(clang)
+        shutil.copyfile(f"{clang}.cmd", clang)
 
-    shutil.copyfile(f"{clang}.cmd", clang)
-
-    arch.check_clang_version()
+    # arch.check_clang_version()
 
 def unpack(ndk: str):
     if is_mac:
         run("7z -aos x -ondk " + ndk)
         run("mv -f ndk/Android\ NDK\ r25c/AndroidNDK9519653.app/Contents/NDK/* ndk/")
         run("rm -rf ndk/Android\ NDK\ r25c")
-    elif is_windows:
+    elif is_windows or is_linux:
         shutil.unpack_archive(ndk, "ndk")
         make_toolchain("ndk/android-ndk-r25c/build/tools/make_standalone_toolchain.py", Arch.arm64)
